% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_survival_models.R
\name{fit_survival_models}
\alias{fit_survival_models}
\title{Fit parametric survival models by treatment arm}
\usage{
fit_survival_models(
  data,
  dists_to_fit = c("exp", "weibull", "gamma", "lnorm", "llogis", "gompertz", "gengamma"),
  arm_names = NULL
)
}
\arguments{
\item{data}{A data frame containing the variables \code{time}, \code{status},
and \code{treat}. \code{status} should be coded as 0 for censored observations
and 1 for events.}

\item{dists_to_fit}{A character vector containing the names of the
parametric distributions to fit.}

\item{arm_names}{An optional named character vector used to label the
treatment arms. The names should correspond to the values of \code{treat}.}
}
\value{
A named list containing the successfully fitted survival models
for each treatment arm.
}
\description{
Fits one or more standard parametric survival distributions separately
for each treatment arm using \code{flexsurv::flexsurvreg()}.
}
\examples{
lung_data <- data.frame(
  time = survival::lung$time,
  status = survival::lung$status - 1,
  treat = survival::lung$sex
)

fits <- fit_survival_models(
  data = lung_data,
  dists_to_fit = c("weibull", "lnorm"),
  arm_names = c(
    "1" = "Treatment 1",
    "2" = "Treatment 2"
  )
)
}
