% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survextrapsimp-package.R
\docType{package}
\name{survextrapsimp-package}
\alias{survextrapsimp}
\alias{survextrapsimp-package}
\title{survextrapsimp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
